+++ kodilla-basic-tests/src/main/java/com/kodilla/basic_assertion/ResultChecker.java
@@ -1,7 +1,7 @@
 package com.kodilla.basic_assertion;
 
 public class ResultChecker {
-    public static boolean assertEquals(int expected, int actual) {      //metoda sprawdzająca (wynik oczekiwany kontra wynik uzyskany
+    public static boolean assertEquals(double expected, double actual) {      //metoda sprawdzająca (wynik oczekiwany kontra wynik uzyskany
         return expected == actual;                                      // oczekiwany ma się równać sprawdzany
     }
 }
